Created ikalog/ui/options.py file.
* No behavior is changed by this PR.
* Moved ikalog.ui.gui.OptionsGUI and ikalog.ui.panel.options.OptionsPanel to
  the file.
* Integrated OptionsPanel into OptionsGUI.
